requirements.txt
streamlit
pandas
python-dotenv
langchain-google-genai
langchain-community
langchain-core